+++ source/App/source/WebApp/Middleware/UserMiddleware.cs
@@ -12,15 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
-using System.Collections.Generic;
 using System.IdentityModel.Tokens.Jwt;
-using System.Linq;
 using System.Net;
 using System.Security.Claims;
-using System.Threading.Tasks;
-using Energinet.DataHub.Core.App.Common;
 using Energinet.DataHub.Core.App.Common.Abstractions.Users;
+using Energinet.DataHub.Core.App.Common.Users;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 
@@ -71,7 +67,7 @@ public sealed class UserMiddleware<TUser> : IMiddleware
             .ProvideUserAsync(userId, actorId, multiTenancy, claimsPrincipal.Claims)
             .ConfigureAwait(false);
 
-        // Domain did not accept the user; returns 401.
+        // Subsystem did not accept the user; returns 401.
         if (user == null)
         {
             context.Response.StatusCode = (int)HttpStatusCode.Unauthorized;