+++ source/TestCommon/documents/release-notes/release-notes.md
@@ -1,5 +1,15 @@
 # TestCommon Release notes
 
+## Version 5.1.1
+
+- Updated Azurite test certificate.
+
+## Version 5.1.0
+
+- Added extensions to class `FunctionAppHostManagerExtensions`:
+    - Method `CheckIfFunctionThrewException`
+    - Method `AssertFunctionWasExecutedAsync`
+
 ## Version 5.0.1
 
 - No functional change